+++ web-ui/src/components/root-components/AppSidebar.tsx
@@ -86,7 +86,7 @@ export const AppSidebar = ({ isMobile, setMobileOpen }: AppSidebarProps) => {
             © 2025 Adamax-Works
           </Typography>
           <Typography variant="caption" display="block" color="text.secondary">
-            v3.5
+            v3.6
           </Typography>
         </Box>
       </Box>