+++ CHANGELOG.md
@@ -362,4 +362,5 @@
 - removed unused code for records
 - added new 'Hall of Fame' page with three tabs:
   - Chronology: displays the list of champions in various leagues and list of grandmasters
+  - Titles: displays the count of championships, wins, podiums etc.
 - made round titles on standings table centered